Consolidate MessageObject into ChatMessage
