+++ Client/src/main/java/com/client/ClientNetworkConnection.java
@@ -14,9 +14,11 @@ import static com.controller.ClientController.QUIT_TYPE_KICK;
 
 public abstract class ClientNetworkConnection {
     private final String notif = " *** ";// notification
-
     private final ListenFromServer connectionThread = new ListenFromServer();
     private final Consumer<Serializable> onReceiveCallback;
+    public ClientNetworkConnection() {
+        this.onReceiveCallback = System.out::println;
+    }
 
     public ClientNetworkConnection(Consumer<Serializable> onReceiveCallback) {
         this.onReceiveCallback = onReceiveCallback;
@@ -38,9 +40,9 @@ public abstract class ClientNetworkConnection {
         connectionThread.quitConnection(QUIT_TYPE_DEFAULT);
     }
 
-    protected abstract String getIP();
+    protected abstract String getServerName();
 
-    protected abstract int getPort();
+    protected abstract int getServerPort();
 
     protected abstract String getUserName();
 
@@ -55,7 +57,7 @@ public abstract class ClientNetworkConnection {
 
         @Override
         public void run() {
-            try (var socket = new Socket(getIP(), getPort());
+            try (var socket = new Socket(getServerName(), getServerPort());
                  var in = new ObjectInputStream(socket.getInputStream());
                  var out = new ObjectOutputStream(socket.getOutputStream())) {
