+++ server.properties
@@ -1,5 +1,5 @@
 #Amir Chat v1.0
-#Fri May 27 08:56:20 EDT 2022
+#Fri May 27 09:52:43 EDT 2022
 MaximumGuest=50
 PortNumber=1436
 roomlist=General;Teen;Music;Party;
